添加了更多反馈；修复了无法保存共有棋子的bug

+++ src/Pages/explorer_decoration/DecorationItem.jsx
@@ -120,7 +120,9 @@ function DecorationItem({ decoration, onEditDecoration, onDeleteDecoration, onPr
                 className="menu-item delete"
                 onClick={() => {
                   setMenuOpen(false)
-                  onDeleteDecoration(decoration.id)
+                  if (window.confirm('确定要删除这个装饰吗？')) {
+                    onDeleteDecoration(decoration.id)
+                  }
                 }}
               >
                 删除
